Load graphics under I/O and KERNAL
Reworks the way graphics are done so that instead of being included in
the final .prg, they are loaded from disk by the code. This allows them
to be placed in the RAM that is underneath the I/O and KERNAL ROM, which
would normally be difficult to use.

In the few places where the graphics memory does need to be accessed,
The I/O and KERNAL memory is temporarily disabled so the underlying RAM
can be seen

This requires that the linker script be modified to not put the graphics
in the final .prg, and that the be extracted separately by objdump and
included in the disk image.

Finally, since cbm.h is used for the routines to load the data from
disk, use it for the COLOR_ defines, since otherwise they conflict