mobs: Delay on update callback
Only call the on_update callback if no on_target_reached callback was
called this frame. This makes the frame time more consistent, at the
expense of delaying the callbacks to be less predicable. In order for
the callbacks to still have accurate timing, the number of elapsed
frames is passed as an argument