Track player and mob position relative to map
Instead of tracking the position of the player and mobs in the global
(sprite) coordinates, track them relative to the map. In addition, use
the center of the sprite as the coordinates instead of the upper left
corner. This changes makes it significantly easier to track which quad
the player of mob is in, since it is now a simple power of 2 division
(right shift). In turn, this makes many calculations easier since the
position in map quads can be expressed in 8 bits for both X and Y
coordinates.