Implement MOB multiplexing
Implements sprite multiplexing for MOBs. If more than 6 MOBs need to be
shown, the code will generate a raster interrupt to write the VICII
registers for the next sprite. This allows more MOBs to be shown,
although there is still a hard limit of 8 sprites per raster line.

The player and sword sprites are never multiplexed