Improve MOB fickerplexing
Makes a slight change to the way that the MOB flickerplexing works.
Instead of removing the MOB with the lowest Y value, offset it by the
number of missed MOBs. This allows the raster interrupts after the first
N mobs to still run as early in the frame as possible.