Remove skipped sprite collision check
Now that sprites are flickerplexed pretty well, remove the code that
tracks when a specific sprite was skipped to prevent collisions with it.
Players should be able to tell where sprites are easily enough since the
no longer completely disappear, and it eliminates a lot of tracking, and
a particularly expensive lookup in the collision checking.