Move some work to the NTSC "dead" frame
A lot of CPU power is wasted in the "dead" frame on NTSC (the extra
frame inserted after 5 to keep the game speed consistent with PAL).
However, there are some calculations that can occur here as they are not
timing dependent. Specifically, the string calculations for the UI, and
the MOB update callbacks (if behind) can be placed here.