Add momentum to the flail
The flail now has momentum where it will start spinning slower and
increase in speed each revolution. Each increase in speed comes with an
increase in damage, and also when the flail hits an enemy it will slow
down instead of stop completely. Only when it hits an enemy while moving
the slowest will it be put away