Allow loading from disk other than 8
Checks the current file device number variable set by the OS to
determine which device to load the graphics from. This means that the
graphics will be loaded from the same device that the user specified in
the "load" command to start the program allowing disks other than 8 to
be used.