mobs: Prevent collisions during damage timer
If a mob is hit, prevent subsequent collisions while the damage timer is
running. This prevents the mob from continuing to take damage from
weapons if the weapon still persists after hitting.

This is most apparent with the flail, where it would register multiple
collisions in a row, dissipating all of the flail damage in a single
blow instead of dealing a single strike and allowing the flail to
continue rotating.