Fix lock up when exiting store
When exiting the store, it's possible for the game to lock up. This is
because we need to be very careful when reading the CTRL 1 register of
the VICII; if the 8th raster bit is set when we do a read-modify-write,
it will set the bit for the interrupt which likely means it will never
fire.